refactor: repair function writer
