+++ src/kiroku/node/kiroku_node.cpp
@@ -25,21 +25,9 @@
 #include <fstream>
 #include <iomanip>
 #include "rclcpp/rclcpp.hpp"
-#include "kiroku/writer/writer.hpp"
 #include "kiroku_interfaces/msg/logger.hpp"
 
-using json = nlohmann::json;
 using std::placeholders::_1;
-
-enum LoggerLevel
-{
-  DEBUG,
-  INFO,
-  WARN,
-  ERROR,
-  FATAL
-};
-json json_log = {};
 
 KirokuNode::KirokuNode()
 : Node("kiroku_node")
@@ -51,37 +39,7 @@ KirokuNode::KirokuNode()
         this->get_logger(), "filename: '%s'\nmessage log: '%s'\nlevel: '%s'\ntime: '%s'",
         msg->filename, msg->message_logger, msg->level, msg->time);
 
-      int logger_level = 0;
-      if (msg->level.compare("DEBUG") == 0) {
-        logger_level = 0;
-      } else if (msg->level.compare("INFO") == 0) {
-        logger_level = 1;
-      } else if (msg->level.compare("WARN") == 0) {
-        logger_level = 2;
-      } else if (msg->level.compare("ERROR") == 0) {
-        logger_level = 3;
-      } else if (msg->level.compare("FATAL") == 0) {
-        logger_level = 4;
-      }
-
-      LoggerLevel filter_level = INFO;
-
-      if (logger_level >= filter_level) {
-        std::ofstream file_logger;
-        file_logger.open(
-          "src/kiroku/" + msg->filename + ".log",
-          std::ios_base::trunc | std::ios_base::out);
-
-        json_log += {
-          {"message_log", msg->message_logger},
-          {"level", msg->level},
-          {"time", msg->time}
-        };
-
-        file_logger << json_log.dump();
-
-        file_logger.close();
-      }
+      write.write_to_file(msg->filename, msg->message_logger, msg->level, msg->time);
     }
   );
 }
